Shift remaining files into both services
