+++ backend/question-service/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "cs3219-ay2425s1-project-g46",
+  "version": "1.0.0",
+  "description": "[![Review Assignment Due Date](https://classroom.github.com/assets/deadline-readme-button-22041afd0340ce965d47ae6ef1cefeee28c7c493a6346c4f15d667ab976d596c.svg)](https://classroom.github.com/a/bzPrOe11)\r # CS3219 Project (PeerPrep) - AY2425S1\r ## Group: G46",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "nodemon index.js"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
+    "express": "^4.20.0",
+    "firebase-admin": "^12.4.0",
+    "nodemon": "^3.1.4"
+  }
+}
